// src/lib/api/generated/swagger-model/githubComYourUsernameBackendProjectInternalServiceUpdateMovieInput.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Series API
 * Backend API for users and movies with MongoDB.
 * OpenAPI spec version: 1.0
 */
import type { GithubComYourUsernameBackendProjectInternalRepositoryMediaData } from './githubComYourUsernameBackendProjectInternalRepositoryMediaData';

export interface GithubComYourUsernameBackendProjectInternalServiceUpdateMovieInput {
  description?: string;
  image?: string;
  media?: GithubComYourUsernameBackendProjectInternalRepositoryMediaData[];
  title?: string;
  type?: string[];
  uploadBy?: string;
}