// src/lib/api/generated/swagger-model/githubComYourUsernameBackendProjectInternalServiceUpdateUserInput.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Series API
 * Backend API for users and movies with MongoDB.
 * OpenAPI spec version: 1.0
 */

export interface GithubComYourUsernameBackendProjectInternalServiceUpdateUserInput {
  email?: string;
  expiredDate?: string;
  isActive?: boolean;
  isAdmin?: boolean;
  name?: string;
  password?: string;
  profile?: string;
}